requirements.txt
# ACIE Python Dependencies
# Compatible version constraints

# Core ML
torch>=2.0.0
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0

# Scientific Computing (relaxed constraints)
numpy>=1.21.0,<2.0
pandas>=1.5.0
scipy>=1.9.0
networkx>=2.8

# Probabilistic Programming
pyro-ppl>=1.8.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.12.0

# Development (optional)
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.3.0
